+++ EssentialApp/EssentialApp/FeedImageDataLoaderWithFallbackComposite.swift
@@ -1,0 +1,42 @@
+//
+//  FeedImageDataLoaderWithFallbackComposite.swift
+//  EssentialApp
+//
+//  Created by Vadym Bohdan on 31.05.2023.
+//
+
+import Foundation
+import EssentialFeed
+
+public class FeedImageDataLoaderWithFallbackComposite: FeedImageDataLoader {
+    private let primary: FeedImageDataLoader
+    private let fallback: FeedImageDataLoader
+    
+    public init(primary: FeedImageDataLoader, fallback: FeedImageDataLoader) {
+        self.primary = primary
+        self.fallback = fallback
+    }
+    
+    private class TaskWrapper: FeedImageDataLoaderTask {
+        var wrapped: FeedImageDataLoaderTask?
+        
+        func cancel() {
+            wrapped?.cancel()
+        }
+    }
+    
+    public func loadImageData(from url: URL, completion: @escaping (FeedImageDataLoader.Result) -> Void) -> FeedImageDataLoaderTask {
+        let task = TaskWrapper()
+        task.wrapped = primary.loadImageData(from: url) { [weak self] result in
+            switch result {
+            case .success:
+                completion(result)
+                
+            case .failure:
+                task.wrapped = self?.fallback.loadImageData(from: url, completion: completion)
+            }
+            
+        }
+        return task
+    }
+}
